Extract focused CLI motion and compaction test modules
Keep existing Hspec group registration order and all example bodies. Share small agent and response fixtures while separating motion, manual OpenAI compaction, and task-plan contracts.
